Reclassify 23 <common> functions via Ghidra decompilation evidence
Reduces <common> from 28 to 7 functions using:
- __FILE__ assert strings for 7 functions (certain)
- Address-span and semantic analysis for 16 functions

New objects: transport_address.obj, xbox_sound_cache.obj,
             message_header.obj, data_packet_groups.obj

Moves by evidence type:
  __FILE__ confirmed: actors.obj (×2), transport_address.obj,
    game_engine.obj, network_server_manager.obj, xbox_sound_cache.obj (×2),
    message_header.obj, data_packet_groups.obj
  Address-span/semantic: cseries.obj, observer.obj, contrails.obj,
    particles.obj, point_physics.obj, main.obj (×4),
    collision_usage.obj (×2), vector_math.obj (×3)

Also fixes: contrails_random_short → FUN_00097c80 in particles.c
            contrail_set_state_for_object usage in objects.c
            game_looping_sound_delete now ported in game_sound.c

+++ src/halo/ai/actor_looking.c
@@ -55,7 +55,7 @@ void FUN_00014540(int actor_handle)
   if (look_entry == -1) {
     /* No existing look entry: use object's world position (type 3) */
     look_buf[0] = 3;
-    FUN_001a9200(look_object, (int *)&look_buf[2]);
+    FUN_001a9200(look_object, (float *)&look_buf[2]);
   } else {
     /* Existing look entry found: use it as an object-handle target (type 1) */
     look_buf[0] = 1;
